standardize on json-schema gem instead of json_schema since the former has better support for $ref resolution. refactor specs to use JSON::Validator
