+++ Gemfile
@@ -17,7 +17,8 @@ gem 'coffee-rails', '~> 4.1.0'
 gem 'email_validator'
 gem 'flutie'
 gem 'jquery-rails'
-gem 'json_schema'
+gem 'json-patch'
+gem 'json-schema'
 gem 'neat', '~> 1.7.0'
 gem 'normalize-rails', '~> 3.0.0'
 gem 'pg'
@@ -31,7 +32,6 @@ gem 'simple_form'
 gem 'sinatra', require: nil
 gem 'title'
 gem 'uglifier'
-gem 'json-patch'
 gem 'redcarpet'
 gem 'savon-multipart', git: 'https://github.com/jessieay/savon-multipart.git'
 gem 'swagger-ui_rails', git: 'https://github.com/pkarman/swagger-ui_rails.git', submodules: true
