+++ spec/docs/examples_spec.rb
@@ -1,4 +1,5 @@
 require 'rails_helper'
+include JsonSchemaSpecHelper
 
 describe 'JSON schemas' do
   before(:all) do
@@ -29,19 +30,4 @@ describe 'JSON schemas' do
     errors = JSON::Validator.fully_validate(schema_file, example, errors_as_objects: true)
     expect(errors).to eq []
   end
-
-  def register_schemas_by_uri
-    schema_files = Dir.glob("#{schema_directory}/*.json")
-    schema_files.each do |schema_file|
-      schema_buf = JSON.parse(File.read(schema_file))
-      schema = JSON::Schema.new(schema_buf, :ignored)
-      JSON::Validator.add_schema(schema)
-    end
-  end
-
-  def schema_directory
-    "#{Rails.public_path}/schemas"
-  end    
 end
-
-
